nyoping-bot-postgres-render/requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
asyncpg>=0.29.0
tzdata>=2024.1
fastapi>=0.110.0
uvicorn>=0.27.0
Jinja2>=3.1.0
requests>=2.31.0
python-multipart>=0.0.9
itsdangerous>=2.1.2
